move script call into a helper
