+++ cli/cli/commands/service/stop/stop.go
@@ -2,10 +2,6 @@ package stop
 
 import (
 	"context"
-	"fmt"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/enclaves"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/services"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/starlark_run_config"
 	"github.com/kurtosis-tech/kurtosis/api/golang/engine/kurtosis_engine_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/highlevel/enclave_id_arg"
@@ -14,6 +10,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/args"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_str_consts"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/shared_starlark_calls"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface"
 	metrics_client "github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/client"
 	"github.com/kurtosis-tech/stacktrace"
@@ -31,11 +28,6 @@ const (
 
 	kurtosisBackendCtxKey = "kurtosis-backend"
 	engineClientCtxKey    = "engine-client"
-
-	starlarkScript = `
-def run(plan, args):
-	plan.stop_service(name=args["service_name"])
-`
 )
 
 var ServiceStopCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosisCommand{
@@ -99,27 +91,9 @@ func run(
 
 		serviceName := serviceContext.GetServiceName()
 
-		if err := StopServiceStarlarkCommand(ctx, enclaveCtx, serviceName); err != nil {
+		if err := shared_starlark_calls.StopServiceStarlarkCommand(ctx, enclaveCtx, serviceName); err != nil {
 			return stacktrace.Propagate(err, "An error occurred stopping service '%v' from enclave '%v'", serviceIdentifier, enclaveIdentifier)
 		}
-	}
-	return nil
-}
-
-func StopServiceStarlarkCommand(ctx context.Context, enclaveCtx *enclaves.EnclaveContext, serviceName services.ServiceName) error {
-	params := fmt.Sprintf(`{"service_name": "%s"}`, serviceName)
-	runResult, err := enclaveCtx.RunStarlarkScriptBlocking(ctx, starlarkScript, starlark_run_config.NewRunStarlarkConfig(starlark_run_config.WithSerializedParams(params)))
-	if err != nil {
-		return stacktrace.Propagate(err, "An unexpected error occurred on Starlark for stopping service")
-	}
-	if runResult.ExecutionError != nil {
-		return stacktrace.NewError("An error occurred during Starlark script execution for stopping service: %s", runResult.ExecutionError.GetErrorMessage())
-	}
-	if runResult.InterpretationError != nil {
-		return stacktrace.NewError("An error occurred during Starlark script interpretation for stopping service: %s", runResult.InterpretationError.GetErrorMessage())
-	}
-	if len(runResult.ValidationErrors) > 0 {
-		return stacktrace.NewError("An error occurred during Starlark script validation for stopping service: %v", runResult.ValidationErrors)
 	}
 	return nil
 }
